Split the tail into document surgery and the round trip
`resolve/tail.py` fused two things behind the largest function in the repo:
pure OTIO document surgery, which needs no Resolve at all, and the
export/import round trip, which is nothing but Resolve. The seam between them
was a `verify` callback the build filled with a lambda over its own private
read-back, so the callee called back into its caller's privates mid-round-trip
and the caller had to know the callback's exception contract and that staging
was deleted only once verify passed.

The surgery moves to `cut/otio.py` — `inject`, `transitions`, and the span and
rate arithmetic under both — with no Resolve import and tests that are plain
dicts rather than fakes.

`materialise` keeps the round trip and loses the callback. Its five context
parameters become one `Staging` value, and it returns a `Landed`: the imported
cut plus the two things still to be done about it. The build runs its own
read-back on `landed.timeline` and then calls `release`, or calls `refuse` and
raises what comes back. Check-before-delete is now the returned object's
contract rather than a callback protocol — a caller that does neither leaves
both timelines standing, which is the safe end of the failure.

Closes #221.
